fix & refactor : Todo 완료 시 오늘의 미션과 하단 주간 목록이 동기화 되지 않는 문제 해결
