+++ src/composite/home/index.ts
@@ -1,5 +1,4 @@
-export { TodayMissionBoard } from './todayMissionBoard/component';
 export { CheerMessageCard } from './cheerMessageCard/component';
-export { WeeklyPlanBoard } from './weeklyPlanBoard/component';
+export { PlanBoard } from './planBoard/component';
 export { GoalRoadMap } from './goalRoadMap/component';
 export { ContributionGraph } from './contributionGraph/component';
