+++ src/feature/todo/index.ts
@@ -1,7 +1,7 @@
 import { PlanSelectorProvider } from '@/feature/todo/PlanSelector/hooks';
 import { PlanSelect } from '@/feature/todo/PlanSelector/component';
 
-export { TodayMissionItem } from './TodayMissionItem/component';
+export { TodayMissionBoard } from './todayMissionBoard/component';
 export { PlanSelect } from './PlanSelector/component';
 export { WeeklyTodoList } from './WeeklyTodoList/ui/WeeklyTodoList';
