+++ src/feature/todo/todayMissionBoard/hooks.ts
@@ -1,5 +1,5 @@
 import { useState, useEffect, useCallback } from 'react';
-import { getTodayMissionList } from './api';
+import { getTodayMissionList, patchTodoStatus } from './api';
 import { Todo } from '@/shared/type/Todo';
 
 export function useTodayMissionList() {
@@ -21,14 +21,38 @@ export function useTodayMissionList() {
     }
   }, []);
 
+  const refetch = useCallback(async () => {
+    await fetchTodayMissionList();
+  }, [fetchTodayMissionList]);
+
   useEffect(() => {
     fetchTodayMissionList();
   }, [fetchTodayMissionList]);
 
   return {
-    data,
+    todayTodoList: data,
     isLoading,
     error,
     fetchTodayMissionList,
+    refetch,
   };
+}
+
+export function usePatchTodoStatus() {
+  const [isLoading, setIsLoading] = useState(false);
+  const [error, setError] = useState<string | null>(null);
+
+  const mutate = useCallback(async (todoId: string, isCompleted: boolean) => {
+    setIsLoading(true);
+    setError(null);
+    try {
+      await patchTodoStatus({ todoId, isCompleted });
+    } catch (err: any) {
+      setError(err?.message || '체크 상태를 변경하지 못했습니다.');
+    } finally {
+      setIsLoading(false);
+    }
+  }, []);
+
+  return { mutate, isLoading, error };
 }
